# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface OrderItem {
  id: string;
  name: string;
  qty: number;
  price: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  baru: "baru",
  disiapkan: "disiapkan",
  selesai: "selesai",
} as const;

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod];

export const PaymentMethod = {
  CASH: "CASH",
  QRIS: "QRIS",
  DEBIT: "DEBIT",
  VA_BCA: "VA_BCA",
  VA_MANDIRI: "VA_MANDIRI",
} as const;

export interface Order {
  id: number;
  orderCode: string;
  tableNumber: string;
  items: OrderItem[];
  subtotal: number;
  total: number;
  paymentMethod: PaymentMethod;
  cashReceived?: number | null;
  status: OrderStatus;
  createdAt: string;
  updatedAt: string;
}

export interface CreateOrderInput {
  tableNumber: string;
  items: OrderItem[];
  subtotal: number;
  total: number;
  paymentMethod: PaymentMethod;
  cashReceived?: number | null;
}

export interface UpdateOrderStatusInput {
  status: OrderStatus;
}

export interface OrdersSummary {
  newOrdersToday: number;
  revenueToday: number;
  totalOrders: number;
}