# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: orders
    description: Snack-bar order operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: clearOrders
      tags: [orders]
      summary: Delete all orders (staff reset)
      responses:
        "204":
          description: All orders cleared
  /orders/summary:
    get:
      operationId: getOrdersSummary
      tags: [orders]
      summary: Today's order summary for staff dashboard
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersSummary"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusInput"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    OrderItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        qty:
          type: integer
        price:
          type: integer
      required: [id, name, qty, price]
    OrderStatus:
      type: string
      enum: [baru, disiapkan, selesai]
    PaymentMethod:
      type: string
      # VA values remain readable for historical orders; new checkout only offers
      # CASH, QRIS, and DEBIT.
      enum: [CASH, QRIS, DEBIT, VA_BCA, VA_MANDIRI]
    Order:
      type: object
      properties:
        id:
          type: integer
        orderCode:
          type: string
        tableNumber:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: integer
        total:
          type: integer
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethod"
        cashReceived:
          type: integer
          nullable: true
        status:
          $ref: "#/components/schemas/OrderStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - orderCode
        - tableNumber
        - items
        - subtotal
        - total
        - paymentMethod
        - status
        - createdAt
        - updatedAt
    CreateOrderInput:
      type: object
      properties:
        tableNumber:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: integer
        total:
          type: integer
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethod"
        cashReceived:
          type: integer
          nullable: true
      required: [tableNumber, items, subtotal, total, paymentMethod]
    UpdateOrderStatusInput:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/OrderStatus"
      required: [status]
    OrdersSummary:
      type: object
      properties:
        newOrdersToday:
          type: integer
        revenueToday:
          type: integer
        totalOrders:
          type: integer
      required: [newOrdersToday, revenueToday, totalOrders]