# lib/api-zod/src/generated/types/createOrderInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { PaymentMethod } from "./paymentMethod";

export interface CreateOrderInput {
  tableNumber: string;
  items: OrderItem[];
  subtotal: number;
  total: number;
  paymentMethod: PaymentMethod;
  cashReceived?: number | null;
}