# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderStatus } from "./orderStatus";
import type { PaymentMethod } from "./paymentMethod";

export interface Order {
  id: number;
  orderCode: string;
  tableNumber: string;
  items: OrderItem[];
  subtotal: number;
  total: number;
  paymentMethod: PaymentMethod;
  cashReceived?: number | null;
  status: OrderStatus;
  createdAt: Date;
  updatedAt: Date;
}