# lib/api-zod/src/generated/types/paymentMethod.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod];

export const PaymentMethod = {
  CASH: "CASH",
  QRIS: "QRIS",
  DEBIT: "DEBIT",
  VA_BCA: "VA_BCA",
  VA_MANDIRI: "VA_MANDIRI",
} as const;